Preparation for 0.1.0-beta.1 (19)

+++ scirs2-core/src/neural_architecture_search.rs
@@ -1531,7 +1531,7 @@ impl NeuralArchitectureSearch {
                 };
 
                 let projection_flops = sequence_length as u64 * hidden_dim * hidden_dim * 2;
-                (num_heads * single_head_flops) + projection_flops as u64
+                (num_heads * single_head_flops) + projection_flops
             }
 
             LayerType::BatchNorm => {
